// src/interfaces/params/companyPaymentParams.interface.ts
export interface IGetCompanyPaymentParams {
    uuid_company: string;
    uuid_company_payment: string;
}

export interface ICreateCompanyPaymentParams {
    uuid_company: string;
}

export interface IUpdateCompanyPaymentParams {
    uuid_company: string;
    uuid_company_payment: string;
}

export interface IDeleteCompanyPaymentParams {
    uuid_company: string;
    uuid_company_payment: string;
}